% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitBessitPG.R
\name{logitBessitPG}
\alias{logitBessitPG}
\title{Logit Model with Independent Bessel Priors: Polya-Gamma Gibbs Sampler}
\usage{
logitBessitPG(nsim, X, y, nu, mu, Sigma, x0 = NULL, burn = 1,
              thin = 1, verbose = +Inf)
}
\arguments{
\item{nsim}{The number of simulations.}

\item{X}{The design matrix.}

\item{y}{The response vector.}

\item{nu}{The shape parameter.}

\item{mu}{The prior location vector.}

\item{Sigma}{The prior scale matrix, it must be diagonal.}

\item{x0}{The starting point.}

\item{burn}{The number of draws to be discarded as burn-in.}

\item{thin}{The thinning parameter.}

\item{verbose}{The period for printing status of the chain.}
}
\value{
Posterior sample of the parameters.
}
\description{
Polya-Gamma Gibbs sampler for the logit model with independent Bessel priors.
It is used the code from package \pkg{BayesLogit} with slight
modifications.
}
\details{
\code{nu} is degrees of freedom for Bessel distributions. If \code{nu == 1}
then the priors are Laplace distributions. If \code{nu} goes to infinity
then the priors converge to Gaussian distributions.

\code{x0}: the starting point is generated in the following way:
\itemize{
\item if \code{x0 == NULL} then the starting point for \code{beta} is the
null vector,
\item \code{else} set the starting point for \code{beta} equal to \code{x0}.
}

Only one value every \code{thin} values is kept in the chain, so the true
number of complete scans will be \code{nsim * thin + burn}. By default
\code{thin = 1}, that is no thinning.

The current time and the current number of iteration are printed one every
\code{verbose} iterations. Furthermore:
\itemize{
\item if \code{verbose == +-Inf} then there is no printing,
\item if \code{verbose != +-Inf} then at least start and end of simulation
are reported.
}

The algorithm described in Polson et al. (2013) is used. The mixture density
is a \eqn{\mathrm{gamma} \big( (\nu + 1)/2, \nu/2 \big)}.

Pure R implementation of package \pkg{BayesLogit} is used for sampling from
Polya-Gamma distribution.
}
\references{
Polson, N. G., Scott, J. G., & Windle, J. (2013). Bayesian
inference for logistic models using Pólya–Gamma latent variables. Journal
of the American statistical Association, 108(504), 1339-1349.
}
