% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitDirLaplPG.R
\name{logitDirLaplPG}
\alias{logitDirLaplPG}
\title{Logit Model with Dirichlet-Laplace Prior: Polya-Gamma Gibbs Sampler}
\usage{
logitDirLaplPG = function(nsim, X, y, mu, Sigma, x0 = NULL, burn = 1,
                          thin = 1, verbose = +Inf)
}
\arguments{
\item{nsim}{The number of simulations.}

\item{X}{The design matrix.}

\item{y}{The response vector.}

\item{mu}{The prior location vector.}

\item{Sigma}{The prior scale matrix, it must be diagonal.}

\item{x0}{The starting point.}

\item{burn}{The number of draws to be discarded as burn-in.}

\item{thin}{The thinning parameter.}

\item{verbose}{The period for printing status of the chain.}
}
\value{
Posterior sample of the parameters.
}
\description{
Polya-Gamma Gibbs sampler for the logit model with Dirichlet-Laplace prior as
described in Bhattacharya at al. (2015). It is used the code from package
\pkg{BayesLogit} with slight modifications.
}
\details{
\code{x0}: the starting point is generated in the following way:
\itemize{
\item if \code{x0 == NULL} then sample \code{x0} from unit circle. Thus set
the starting value of \code{beta} equal to \code{x0} times the square root
of the expected value of the scaling random variable,
\item \code{else} set the starting point for \code{beta} equal to \code{x0}.
}
Only one value every \code{thin} values is kept in the chain, so the true
number of complete scans will be \code{nsim * thin + burn}. By default
\code{thin = 1}, that is no thinning.

The current time and the current number of iteration are printed one every
\code{verbose} iterations. Furthermore:
\itemize{
\item if \code{verbose == +-Inf} then there is no printing,
\item if \code{verbose != +-Inf} then at least start and end of simulation
are reported.
}

The algorithm described in Polson et al. (2013) is used. A uniform discrete
prior on the support \eqn{ \{ \frac{1}{300}, \frac{2}{300}, \dots,
 \frac{299}{300}, 1 \} } is used for the parameter of the Dirichlet-Laplace
prior and, in the Gibbs sampler, its starting points is always \eqn{0.5}.

Pure R implementation of package \pkg{BayesLogit} is used for sampling from
Polya-Gamma distribution.
}
\references{
Bhattacharya, A., Pati, D., Pillai, N. S., & Dunson, D. B. (2015).
Dirichlet–Laplace priors for optimal shrinkage. Journal of the American
Statistical Association, 110(512), 1479-1490.

Polson, N. G., Scott, J. G., & Windle, J. (2013). Bayesian
inference for logistic models using Pólya–Gamma latent variables. Journal
of the American statistical Association, 108(504), 1339-1349.
}
