% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitGaussPG.R
\name{logitGaussPG}
\alias{logitGaussPG}
\title{Logit Model with Gaussian Prior: Polya-Gamma Gibbs Sampler}
\usage{
logitGaussPG = function(nsim, X, y, mu, Sigma, x0 = NULL, burn = 1,
                         thin = 1, verbose = +Inf)
}
\arguments{
\item{nsim}{The number of simulations.}

\item{X}{The design matrix.}

\item{y}{The response vector.}

\item{mu}{The prior location vector.}

\item{Sigma}{The prior scale matrix.}

\item{x0}{The starting point.}

\item{burn}{The number of draws to be discarded as burn-in.}

\item{thin}{The thinning parameter.}

\item{verbose}{The period for printing status of the chain.}
}
\value{
Posterior sample of the parameters.
}
\description{
Polya-Gamma Gibbs sampler for the logit model with Gaussian prior. It is
used the code from package \pkg{BayesLogit} with slight modifications.
}
\details{
\code{x0}: the starting point is generated in the following way:
\itemize{
\item if \code{x0 == NULL} then the starting point for \code{beta} is the
null vector,
\item \code{else} set the starting point for \code{beta} equal to \code{x0}.
}

Only one value every \code{thin} values is kept in the chain, so the true
number of complete scans will be \code{nsim * thin + burn}. By default
\code{thin = 1}, that is no thinning.

The current time and the current number of iteration are printed one every
\code{verbose} iterations. Furthermore:
\itemize{
\item if \code{verbose == +-Inf} then there is no printing,
\item if \code{verbose != +-Inf} then at least start and end of simulation
are reported.
}

The algorithm described in Polson et al. (2013) is used.

Pure R implementation of package \pkg{BayesLogit} is used for sampling from
Polya-Gamma distribution.
}
\references{
Polson, N. G., Scott, J. G., & Windle, J. (2013). Bayesian
inference for logistic models using Pólya–Gamma latent variables. Journal
of the American statistical Association, 108(504), 1339-1349.
}
