% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitStudentUPG.R
\name{logitStudentUPG}
\alias{logitStudentUPG}
\title{Logit Model with Student-t Prior: Ultimate Polya-Gamma Gibbs Sampler}
\usage{
logitStudentUPG = function(nsim, X, y, nu, Sigma, x0 = NULL, burn = 1,
                           thin = 1, verbose = +Inf)
}
\arguments{
\item{nsim}{The number of simulations.}

\item{X}{The design matrix.}

\item{y}{The response vector.}

\item{nu}{The shape parameter.}

\item{Sigma}{The prior scale matrix, it must be diagonal.}

\item{x0}{The starting point.}

\item{burn}{The number of draws to be discarded as burn-in.}

\item{thin}{The thinning parameter.}

\item{verbose}{The period for printing status of the chain.}
}
\value{
Posterior sample of the parameters.
}
\description{
Ultimate Polya-Gamma Gibbs sampler for the logit model with Student-t prior.
It is used the code from package \pkg{UPG} with slight modifications.
}
\details{
\code{x0}: the starting point is generated in the following way:
\itemize{
\item if \code{x0 == NULL} then the starting point for \code{beta} is the
null vector,
\item \code{else} set the starting point for \code{beta} equal to \code{x0}.
}

Only one value every \code{thin} values is kept in the chain, so the true
number of complete scans will be \code{nsim * thin + burn}. By default
\code{thin = 1}, that is no thinning.

The current time and the current number of iteration are printed one every
\code{verbose} iterations. Furthermore:
\itemize{
\item if \code{verbose == +-Inf} then there is no printing,
\item if \code{verbose != +-Inf} then at least start and end of simulation
are reported.
}

The algorithm described in Zens et al. (2023) is used. The mixture density
is a \eqn{\mathrm{inv.gamma} \big( \nu/2, \nu/2 \big)}.

Pure R implementation of package \pkg{BayesLogit} is used for sampling from
Polya-Gamma distribution.
}
\references{
Zens, G., Frühwirth-Schnatter, S., & Wagner, H. (2023). Ultimate Pólya Gamma
Samplers–Efficient MCMC for possibly imbalanced binary and categorical
data. Journal of the American Statistical Association, 1-12.
}
