% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSBR_ProbitGauss_VarSel.R
\name{LSBR_ProbitGauss_VarSel}
\alias{LSBR_ProbitGauss_VarSel}
\title{Probit Model with Gaussian Prior: Variables Selection}
\usage{
LSBR_ProbitGauss_VarSel = function(nsim, X, y, mu, Sigma,
                                   ntim = 1e+4, x0 = NULL, verbose = +Inf)
}
\arguments{
\item{nsim}{The number of simulations.}

\item{X}{The design matrix.}

\item{y}{The response vector.}

\item{mu}{The prior location vector.}

\item{Sigma}{The prior scale matrix.}

\item{ntim}{The number of draws for the normalizing constant computation.}

\item{x0}{The starting point.}

\item{verbose}{The period for printing status of the chain.}
}
\value{
Posterior inclusion probabilities and a posterior sample from model
space.
}
\description{
Variable Selection using Bayes Factor and Metropolis steps in a Linear
Symmetric Binary Regression (LSBR) model space.
}
\details{
A posterior sample from the parameter space is drawn and a
Metropolis-within-Gibbs algorithm is used. Given a model, the marginal data
density is computed by using the procedure described in Durante (2019).
The update from the model space is the same as in Onorati and Liseo (2022).

The prior on model space is the discrete uniform on models including the
intercept. Models without the intercept are not included in the space model.
The prior on parameters is a Gaussian distribution with mean vector \code{mu}
and variance-covariance matrix \code{Sigma}.

If \code{x0 == NULL} then the starting point is get using the posterior
credible interval at 0.95 level. The i-th variable is included in the
starting point if zero is not included

The current time and the current number of iteration are printed one every
\code{verbose} iterations. Furthermore:
\itemize{
\item if \code{verbose == +-Inf} then there is no printing,
\item if \code{verbose != +-Inf} then at least start and end of simulation
are reported.
}
}
\references{
Durante, D. (2019). Conjugate Bayes for probit regression via unified
skew-normal distributions. Biometrika, 106(4), 765-779.

Onorati, P., & Liseo, B. (2022). An extension of the unified skew-normal
family of distributions and application to Bayesian binary regression.
arXiv preprint arXiv:2209.03474.
}
