% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmvnctID.R
\name{pmvnctID}
\alias{pmvnctID}
\title{The Multivariate Non-Central Student-t Distribution with Identity
Matrix}
\usage{
pmvnctID = function(x, nu = NULL, delta, n = NULL, dx = NULL, log.p = FALSE,
                    maxComp = 1e+9, stringLength = 100)
}
\arguments{
\item{x}{Matrix of arguments.}

\item{nu}{Shape parameter.}

\item{delta}{Non centrality vector.}

\item{n}{Dimension of multivariate non-central Student-t.}

\item{dx}{Vector of pseudo-sampled points.}

\item{log.p}{If it is true then the logarythm is returned.}

\item{maxComp}{Maximum number of rows of x to be computed for each iteration.}

\item{stringLength}{Number of pseudo-sapled points to be computed. It is used
only if the points are not provided.}
}
\value{
Vector of distribution function values.
}
\description{
Distribution function for the multivariate non-central Student-t distribution
with identity matrix as scale matrix.
}
\details{
It is said \eqn{X} follows a multivariate non-central Student-t distribution
with shape \eqn{\nu}, scale matrix \eqn{\Omega} and non centrality vector
\eqn{\delta} if
\deqn{
   X = \sqrt{W} (Z + \delta) \, ,
 }
where \eqn{Z \sim N(0, \Omega) \perp\!\!\!\perp W \sim \mathrm{inv.gamma}
 (\nu / 2, \nu / 2)}. Notice that this function requires \eqn{\Omega = I}.

The distribution function is computed via a quasi Monte Carlo algorithm as
described in Onorati and Liseo (2022).
}
\references{
Onorati, P., & Liseo, B. (2022). An extension of the unified skew-normal
family of distributions and application to Bayesian binary regression.
arXiv preprint arXiv:2209.03474.
}
