% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSUN_LogitDirLapl_Gibbs.R
\name{pSUN_LogitDirLapl_Gibbs}
\alias{pSUN_LogitDirLapl_Gibbs}
\title{Perturbed Unified Skew Normal Logit Dirichlet-Laplace: Gibbs Sampler}
\usage{
pSUN_LogitDirLapl_Gibbs = function(nsim, A, b, xi, Omega, x0 = NULL,
                                   burn = 1, thin = 1, plim = 1 / nsim,
                                   verbose = +Inf)
}
\arguments{
\item{nsim}{The number of simulations.}

\item{A}{The slant matrix.}

\item{b}{The slant vector.}

\item{xi}{The location vector.}

\item{Omega}{The scale matrix, it must be diagonal.}

\item{x0}{The starting point.}

\item{burn}{The number of draws to be discarded as burn-in.}

\item{thin}{The thinning parameter.}

\item{plim}{Minimum empirical acceptance probability such that the
function does not stop with an error.}

\item{verbose}{The period for printing status of the chain.}
}
\value{
Sample of the pSUN distribution.
}
\description{
Gibbs Sampler for a perturbed unified skew normal (pSUN) logit
Dirichlet-Laplace distribution.
}
\details{
It is said \eqn{X} follows a pSUN logit Dirichlet-Laplace distribution if
\deqn{
  X = \xi + \mathrm{diag}^\frac{1}{2} (\Omega) \, Z \vert T \le A Z + b \\
  Z \vert W \sim N(0, \bar{\Omega}_W) \perp\!\!\!\perp T \vert V \sim N(0,
   \Theta_V) \\
  W_i \overset{i.i.d.}{\sim} \mathrm{SDL}(\cdot; \psi)
   \perp\!\!\!\perp V_i \overset{i.i.d.}{\sim} \mathrm{LK}(\cdot)
 }
where \eqn{\bar{\Omega}_W = \mathrm{diag}^\frac{1}{2}(W) \bar{\Omega}
 \mathrm{diag}^\frac{1}{2}(W)}, \eqn{\Theta_V = \mathrm{diag}^\frac{1}{2}(V)
 \Theta \mathrm{diag}^\frac{1}{2}(V)}, \eqn{\bar{\Omega}} is the correlation
matrix of \eqn{\Omega} and \eqn{\mathrm{LK}(\cdot)} is the CDF of a logistic
Kolmogorov distribution. \eqn{\mathrm{SDL}(\cdot; \psi)} is the CDF of the
scaling mixture distribution of the Dirichlet-Laplace prior with parameter
\eqn{\psi} (Bhattacharya at al., 2015).

In this function, \eqn{\Theta} is restricted to be the identity matrix.
Notice that a logistic Kolmogorov distribution is four times the square of
a Kolmogorov random variable. Furthermore, \eqn{W_i \overset{i.i.d.}{\sim}
 \mathrm{SDL}(\cdot; \psi)} implies
\deqn{
  W_i =  G^2_i E_i \, , \, G_i \overset{i.i.d.}{\sim}
  \mathrm{gamma}(\psi, 1 / 2) \perp\!\!\!\perp E_i \overset{i.i.d.}{\sim}
  \exp(1 / 2)
 }
or, equivalently
\deqn{
  W = \widetilde{G}_i U E \\
  \widetilde{G}_i \overset{i.i.d.}{\sim} \mathrm{gamma}(d \psi, 1 / 2) \\
  U \sim \mathrm{Dirichlet}_d(\psi, \psi, \dots, \psi) \\
  E_i \overset{i.i.d.}{\sim} \exp(1 / 2)
 }
a uniform discrete prior on the support \eqn{ \{ \frac{1}{300},
 \frac{2}{300}, \dots, \frac{299}{300}, 1 \} } is used for \eqn{\psi}.

Draws from truncated Gaussian are obtained using Botev (2017) algorithm. See
Onorati and Liseo (2022) for more details.

\code{x0}: the starting point is generated in the following way:
\itemize{
\item if \code{x0 == NULL} then sample \code{x0} from unit circle. Thus set
the starting value of \code{beta} equal to \code{x0} times the square root
of the expected value of the scaling random variable,
\item \code{else} set the starting point for \code{beta} equal to \code{x0}.
}

Only one value every \code{thin} values is kept in the chain, so the true
number of complete scans will be \code{nsim * thin + burn}. By default
\code{thin = 1}, that is no thinning.

The current time and the current number of iteration are printed one every
\code{verbose} iterations. Furthermore:
\itemize{
\item if \code{verbose == +-Inf} then there is no printing,
\item if \code{verbose != +-Inf} then at least start and end of simulation
are reported.
}
}
\references{
Bhattacharya, A., Pati, D., Pillai, N. S., & Dunson, D. B. (2015).
Dirichlet–Laplace priors for optimal shrinkage. Journal of the American
Statistical Association, 110(512), 1479-1490.

Botev, Z. I. (2017). The normal law under linear restrictions: simulation and
estimation via minimax tilting. Journal of the Royal Statistical Society
Series B: Statistical Methodology, 79(1), 125-148.

Onorati, P., & Liseo, B. (2022). An extension of the unified skew-normal
family of distributions and application to Bayesian binary regression.
arXiv preprint arXiv:2209.03474.
}
