% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSUN_LogitGauss_IS.R
\name{pSUN_LogitGauss_IS}
\alias{pSUN_LogitGauss_IS}
\title{Perturbed Unified Skew Normal Logit Gaussian: Importance Sampling}
\usage{
pSUN_LogitGauss_IS = function(nsim, A, b, xi, Omega, nuSUT = NULL,
                              V = NULL, plim = 1 / nsim)
}
\arguments{
\item{nsim}{The number of simulations.}

\item{A}{The slant matrix.}

\item{b}{The slant vector.}

\item{xi}{The location vector.}

\item{Omega}{The scale matrix.}

\item{nuSUT}{The shape parameter of SUT distribution.}

\item{V}{The vector of constant values for the scales of selective variable.}

\item{plim}{Minimum empirical acceptance probability such that the
function does not stop with an error.}
}
\value{
Sample, normalized weights and logarithm of normalizing constant of
the pSUN distribution.
}
\description{
Importance sampling for a perturbed unified skew normal (pSUN) logit Gaussian
distribution.
}
\details{
It is said \eqn{X} follows a pSUN logit Gaussian distribution if
\deqn{
  X = \xi + \mathrm{diag}^\frac{1}{2} (\Omega) \, Z \vert T \le A Z + b \\
  Z \vert W \sim N(0, \bar{\Omega}_W) \perp\!\!\!\perp T \vert V \sim N(0,
   \Theta_V) \\
  W_i = 1 \, , \, V_i \overset{i.i.d.}{\sim} \mathrm{LK}(\cdot)
 }
where \eqn{\bar{\Omega}_W = \mathrm{diag}^\frac{1}{2}(W) \bar{\Omega}
 \mathrm{diag}^\frac{1}{2}(W)}, \eqn{\Theta_V = \mathrm{diag}^\frac{1}{2}(V)
 \Theta \mathrm{diag}^\frac{1}{2}(V)}, \eqn{\bar{\Omega}} is the correlation
matrix of \eqn{\Omega} and \eqn{\mathrm{LK}(\cdot)} is the CDF of a logistic
Kolmogorov distribution.

Notice that a logistic Kolmogorov distribution is four times the square of
a Kolmogorov random variable.

The importance density is a Unified Skew-t distribution (SUT). It is said
\eqn{X} follows a SUT distribution if
\deqn{
  X = \xi + \sqrt{S} \mathrm{diag}^\frac{1}{2} (\Omega) \, Z  \\
  Z \sim \mathrm{SUN}(\Theta, A, b, 0, \bar{\Omega}) \perp\!\!\!\perp S \sim
  \mathrm{inv.gamma}(\nu / 2, \nu / 2)
 }
where \eqn{\bar{\Omega}} is the correlation matrix of \eqn{\Omega}. In this
function, \eqn{\Theta} is restricted to be the identity matrix.
See Onorati and Liseo (2022) for more details.
}
\references{
Onorati, P., & Liseo, B. (2022). An extension of the unified skew-normal
family of distributions and application to Bayesian binary regression.
arXiv preprint arXiv:2209.03474.
}
