% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSUN_LogitStudent_Gibbs.R
\name{pSUN_LogitStudent_Gibbs}
\alias{pSUN_LogitStudent_Gibbs}
\title{Perturbed Unified Skew Normal Logit Student-t: Gibbs Sampler}
\usage{
pSUN_LogitStudent_Gibbs = function(nsim, A, b, nu, xi, Omega, x0 = NULL,
                                   burn = 1, thin = 1, plim = 1 / nsim,
                                   verbose = +Inf)
}
\arguments{
\item{nsim}{The number of simulations.}

\item{A}{The slant matrix.}

\item{b}{The slant vector.}

\item{nu}{The shape parameter.}

\item{xi}{The location vector.}

\item{Omega}{The scale matrix.}

\item{x0}{The starting point.}

\item{burn}{The number of draws to be discarded as burn-in.}

\item{thin}{The thinning parameter.}

\item{plim}{Minimum empirical acceptance probability such that the
function does not stop with an error.}

\item{verbose}{The period for printing status of the chain.}
}
\value{
Sample of the pSUN distribution.
}
\description{
Gibbs Sampler for a perturbed unified skew normal (pSUN) logit Student-t
distribution.
}
\details{
It is said \eqn{X} follows a pSUN logit Student-t distribution if
\deqn{
  X = \xi + \mathrm{diag}^\frac{1}{2} (\Omega) \, Z \vert T \le A Z + b \\
  Z \vert W \sim N(0, \bar{\Omega}_W) \perp\!\!\!\perp T \vert V \sim N(0,
   \Theta_V) \\
  W_i \overset{i.i.d.}{\sim} \mathrm{inv.gamma}(\nu / 2, \nu / 2)
   \perp\!\!\!\perp V_i \overset{i.i.d.}{\sim} \mathrm{LK}(\cdot)
 }
where \eqn{\bar{\Omega}_W = \mathrm{diag}^\frac{1}{2}(W) \bar{\Omega}
 \mathrm{diag}^\frac{1}{2}(W)}, \eqn{\Theta_V = \mathrm{diag}^\frac{1}{2}(V)
 \Theta \mathrm{diag}^\frac{1}{2}(V)}, \eqn{\bar{\Omega}} is the correlation
matrix of \eqn{\Omega} and \eqn{\mathrm{LK}(\cdot)} is the CDF of a logistic
Kolmogorov distribution.

Notice that a logistic Kolmogorov distribution is four times the square of
a Kolmogorov random variable.

Draws from truncated Gaussian are obtained using Botev (2017) algorithm. See
Onorati and Liseo (2022) for more details.

\code{nu} is degrees of freedom for Student-t distribution. If \code{nu == 1}
then the random variable is a Cauchy distribution. If \code{nu} goes to
infinity then the random variables converge to Gaussian distribution.

\code{x0}: the starting point is generated in the following way:
\itemize{
\item if \code{x0 == NULL} then the starting point for \code{X} is the
null vector,
\item \code{else} set the starting point for \code{X} equal to \code{x0}.
}

Only one value every \code{thin} values is kept in the chain, so the true
number of complete scans will be \code{nsim * thin + burn}. By default
\code{thin = 1}, that is no thinning.

The current time and the current number of iteration are printed one every
\code{verbose} iterations. Furthermore:
\itemize{
\item if \code{verbose == +-Inf} then there is no printing,
\item if \code{verbose != +-Inf} then at least start and end of simulation
are reported.
}
}
\references{
Botev, Z. I. (2017). The normal law under linear restrictions: simulation and
estimation via minimax tilting. Journal of the Royal Statistical Society
Series B: Statistical Methodology, 79(1), 125-148.

Onorati, P., & Liseo, B. (2022). An extension of the unified skew-normal
family of distributions and application to Bayesian binary regression.
arXiv preprint arXiv:2209.03474.
}
