% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSUN_ProbitGauss_RNG.R
\name{pSUN_ProbitGauss_RNG}
\alias{pSUN_ProbitGauss_RNG}
\title{Perturbed Unified Skew Normal Probit Gaussian: Random Number Generator}
\usage{
pSUN_ProbitGauss_RNG = function(nsim, A, b, xi, Omega, plim = 1 / nsim)
}
\arguments{
\item{nsim}{The number of simulations.}

\item{A}{The slant matrix.}

\item{b}{The slant vector.}

\item{xi}{The location vector.}

\item{Omega}{The scale matrix.}

\item{plim}{Minimum empirical acceptance probability such that the
function does not stop with an error.}
}
\value{
Sample and logarithm of normalizing constant of
the pSUN distribution.
}
\description{
Exact random number generator for a perturbed unified skew normal (pSUN)
probit Gaussian distribution. It is actually a SUN distribution
(Arellano-Valle and Azzalini, 2006) with a different parametrization.
}
\details{
It is said \eqn{X} follows a pSUN probit Gaussian distribution if
\deqn{
  X = \xi + \mathrm{diag}^\frac{1}{2} (\Omega) \, Z \vert T \le A Z + b \\
  Z \vert W \sim N(0, \bar{\Omega}_W) \perp\!\!\!\perp T \vert V \sim N(0,
   \Theta_V) \\
  W_i = 1 \, , \, V_i = 1
 }
where \eqn{\bar{\Omega}_W = \mathrm{diag}^\frac{1}{2}(W) \bar{\Omega}
 \mathrm{diag}^\frac{1}{2}(W)}, \eqn{\Theta_V = \mathrm{diag}^\frac{1}{2}(V)
 \Theta \mathrm{diag}^\frac{1}{2}(V)}, \eqn{\bar{\Omega}} is the correlation
matrix of \eqn{\Omega}.

Draws from truncated Gaussian are obtained using Botev (2017) algorithm.
}
\references{
Arellano‐Valle, R. B., & Azzalini, A. (2006). On the unification of families
of skew‐normal distributions. Scandinavian Journal of Statistics, 33(3),
561-574.

Botev, Z. I. (2017). The normal law under linear restrictions: simulation and
estimation via minimax tilting. Journal of the Royal Statistical Society
Series B: Statistical Methodology, 79(1), 125-148.
}
